Some updates to auth PR
