+++ client/src/pages/Landing.js
@@ -1,7 +1,13 @@
 import React from "react";
+import { Redirect } from "react-router-dom";
 import Hero from "../Components/Hero";
 
-function Landing() {
+function Landing(props) {
+
+  if (props.user) {
+    return <Redirect to="/home" />
+  }
+  
   return (
     <div>
       <Hero />
