+++ client/src/pages/Login.js
@@ -1,7 +1,12 @@
 import React from "react";
+import { Redirect } from "react-router-dom";
 import UserLogin from "../Components/UserLogin";
 
 function userLogin(props) {
+
+  if (props.user) {
+    return <Redirect to="/home" />
+  }
   
   return (
     <div>
