+++ client/src/pages/SignupUser.js
@@ -1,8 +1,14 @@
 import React from "react";
+import { Redirect } from "react-router-dom";
 import UserSignUp from "../Components/UserSignUp";
 
 
 function SignupUser(props) {
+
+  if (props.user) {
+    return <Redirect to="/home" />
+  }
+
   return (
     <div>
       <UserSignUp user={props.user} updateUser={props.updateUser}/>
@@ -11,6 +17,3 @@ function SignupUser(props) {
 }
 
 export default SignupUser;
-
-// https://material-ui.com/components/app-bar/
-// https://material-ui.com/getting-started/templates/
